streamlit-web/requirements.txt
streamlit
google-auth
google-auth-oauthlib
google-api-python-client
yt-dlp
ffmpeg-python
pandas